+++ src/onaptests/steps/onboard/service.py
@@ -86,7 +86,8 @@ class ServiceOnboardStep(BaseStep):
     def cleanup(self) -> None:
         """Cleanup service onboard step."""
         service: Service = Service(name=settings.SERVICE_NAME)
-        service.delete()
+        if service.exists():
+            service.delete()
         super().cleanup()
 
 
@@ -265,5 +266,6 @@ class YamlTemplateServiceOnboardStep(YamlTemplateBaseStep):
     def cleanup(self) -> None:
         """Cleanup service onboard step."""
         service: Service = Service(name=self.service_name)
-        service.delete()
+        if service.exists():
+            service.delete()
         super().cleanup()
