+++ src/onaptests/steps/onboard/vendor.py
@@ -32,6 +32,7 @@ class VendorOnboardStep(BaseStep):
     @BaseStep.store_state(cleanup=True)
     def cleanup(self) -> None:
         vendor: Vendor = Vendor(name=settings.VENDOR_NAME)
-        vendor.archive()
-        vendor.delete()
+        if vendor.exists():
+            vendor.archive()
+            vendor.delete()
         super().cleanup()
